chore: format code

+++ test/foundry/AutoRefunder.t.sol
@@ -33,13 +33,7 @@ contract AutoRefunderTest is Test {
         ERC1967Proxy proxy_ = new ERC1967Proxy(
             address(autoRefunderImpl),
             abi.encodeWithSelector(
-                AutoRefunder.initialize.selector,
-                admin,
-                manager,
-                bot,
-                pauser,
-                address(stakeManager),
-                vault
+                AutoRefunder.initialize.selector, admin, manager, bot, pauser, address(stakeManager), vault
             )
         );
         autoRefunder = AutoRefunder(payable(address(proxy_)));
@@ -49,9 +43,7 @@ contract AutoRefunderTest is Test {
         assertEq(autoRefunder.refundRatio(), 6000);
         assertEq(autoRefunder.refundDays(), 30);
 
-        assertTrue(
-            autoRefunder.hasRole(autoRefunder.DEFAULT_ADMIN_ROLE(), admin)
-        );
+        assertTrue(autoRefunder.hasRole(autoRefunder.DEFAULT_ADMIN_ROLE(), admin));
         assertTrue(autoRefunder.hasRole(autoRefunder.MANAGER(), manager));
         assertTrue(autoRefunder.hasRole(autoRefunder.BOT(), bot));
         assertTrue(autoRefunder.hasRole(autoRefunder.PAUSER(), pauser));
@@ -66,32 +58,28 @@ contract AutoRefunderTest is Test {
 
         SLisBNB slisBnbImpl = new SLisBNB();
         TransparentUpgradeableProxy slisBnbProxy = new TransparentUpgradeableProxy(
-                address(slisBnbImpl),
-                proxyAdminOwner,
-                abi.encodeWithSignature("initialize(address)", admin)
-            );
+            address(slisBnbImpl), proxyAdminOwner, abi.encodeWithSignature("initialize(address)", admin)
+        );
         slisBnb = SLisBNB(address(slisBnbProxy));
 
         ListaStakeManager stakeManagerImpl = new ListaStakeManager();
         TransparentUpgradeableProxy stakeManagerProxy = new TransparentUpgradeableProxy(
-                address(stakeManagerImpl),
-                proxyAdminOwner,
-                abi.encodeWithSignature(
-                    "initialize(address,address,address,address,uint256,address,address)",
-                    address(slisBnb),
-                    admin,
-                    manager,
-                    bot,
-                    synFee,
-                    revenuePool,
-                    validator
-                )
-            );
+            address(stakeManagerImpl),
+            proxyAdminOwner,
+            abi.encodeWithSignature(
+                "initialize(address,address,address,address,uint256,address,address)",
+                address(slisBnb),
+                admin,
+                manager,
+                bot,
+                synFee,
+                revenuePool,
+                validator
+            )
+        );
         stakeManager = ListaStakeManager(payable(address(stakeManagerProxy)));
 
-        assertTrue(
-            stakeManager.hasRole(stakeManager.DEFAULT_ADMIN_ROLE(), admin)
-        );
+        assertTrue(stakeManager.hasRole(stakeManager.DEFAULT_ADMIN_ROLE(), admin));
         vm.startPrank(admin);
         slisBnb.setStakeManager(address(stakeManager));
         vm.stopPrank();
@@ -105,15 +93,13 @@ contract AutoRefunderTest is Test {
 
         uint256 totalAmount = 100 ether;
 
-        (bool success, ) = address(autoRefunder).call{value: totalAmount}("");
+        (bool success,) = address(autoRefunder).call{value: totalAmount}("");
         require(success, "Funding failed");
         assertEq(address(autoRefunder).balance, totalAmount);
 
         vm.startPrank(admin);
         stakeManager.grantRole(stakeManager.MANAGER(), address(autoRefunder));
-        assertTrue(
-            stakeManager.hasRole(stakeManager.MANAGER(), address(autoRefunder))
-        );
+        assertTrue(stakeManager.hasRole(stakeManager.MANAGER(), address(autoRefunder)));
         vm.stopPrank();
 
         vm.expectRevert();
@@ -126,14 +112,11 @@ contract AutoRefunderTest is Test {
         assertEq(address(autoRefunder).balance, 0);
         assertEq(address(stakeManager).balance, (totalAmount * 6000) / 10000);
         assertEq(address(vault).balance, (totalAmount * 4000) / 10000);
-        assertEq(
-            IERC20Metadata(address(slisBnb)).balanceOf(address(stakeManager)),
-            (totalAmount * 6000) / 10000
-        );
+        assertEq(IERC20Metadata(address(slisBnb)).balanceOf(address(stakeManager)), (totalAmount * 6000) / 10000);
 
-        uint dailySlisBnb;
-        uint remainingSlisBnb;
-        uint lastBurnTime;
+        uint256 dailySlisBnb;
+        uint256 remainingSlisBnb;
+        uint256 lastBurnTime;
 
         (dailySlisBnb, remainingSlisBnb, lastBurnTime) = stakeManager.refund();
         assertEq(dailySlisBnb, (totalAmount * 6000) / 10000 / 30);
